+++ safire/src/renderer.c
@@ -1,0 +1,4 @@
+#include "../include/safire/pipeline.h"
+
+
+